+++ crates/test-collector/src/lib.rs
@@ -12,7 +12,6 @@ use cairo_lang_defs::plugin::PluginDiagnostic;
 use cairo_lang_diagnostics::ToOption;
 use cairo_lang_filesystem::cfg::{Cfg, CfgSet};
 use cairo_lang_filesystem::db::init_dev_corelib;
-use cairo_lang_filesystem::detect::detect_corelib;
 use cairo_lang_filesystem::ids::CrateId;
 use cairo_lang_lowering::ids::ConcreteFunctionWithBodyId;
 use cairo_lang_project::{DeserializationError, ProjectConfig, ProjectConfigContent};
@@ -286,7 +285,7 @@ pub fn collect_tests(
     output_path: Option<&str>,
     linked_libraries: Option<Vec<LinkedLibrary>>,
     builtins: Option<Vec<&str>>,
-    corelib_path: Option<&str>,
+    corelib_path: PathBuf,
 ) -> Result<(Program, Vec<TestCase>)> {
     // code taken from crates/cairo-lang-test-runner/src/lib.rs
     let db = &mut {
@@ -297,13 +296,7 @@ pub fn collect_tests(
         b.build()?
     };
 
-    init_dev_corelib(
-        db,
-        corelib_path.map_or_else(
-            || detect_corelib().ok_or_else(|| anyhow!("Failed to load development corelib")),
-            |corelib_path| Ok(corelib_path.into()),
-        )?,
-    );
+    init_dev_corelib(db, corelib_path);
 
     let main_crate_ids = setup_project(db, Path::new(&input_path))
         .with_context(|| format!("Failed to setup project for path({input_path})"))?;